Rows and columns of a data frame, chosen as illumex chooses them

The data an illumex function would use after subset and cols: the rows first, then the columns. Row names keep the original row numbers (or the original names, if the data had them), and the attribute "ilm_select" says what was chosen.

Usage

ilm_subset(
  data,
  subset = NULL,
  subset_negate = FALSE,
  subset_fixed = FALSE,
  cols = NULL,
  cols_negate = FALSE,
  cols_fixed = FALSE
)

Arguments

data

A data frame.

subset, subset_negate, subset_fixed

Which rows; see ilm_selection.

cols, cols_negate, cols_fixed

Which columns; see ilm_selection.

Value

A data frame.

See also

Examples

d <- ilm_sim()
ilm_subset(d, subset = c(grp = "^a"), cols = "^score|^income")
#>     score    income
#> 3   43.30  43453.60
#> 4   49.12  62783.47
#> 6   50.94  18491.95
#> 9   51.61  81282.17
#> 11  46.34  21992.48
#> 13  41.51  16933.64
#> 15  40.64  15387.67
#> 16  43.16  14065.43
#> 18  50.92  73051.99
#> 19  43.97  22053.68
#> 20  47.59  11184.50
#> 21  47.60  37621.41
#> 22  57.01   9394.14
#> 23  40.21  18707.47
#> 24  44.99  54965.26
#> 25  57.73  38591.37
#> 26  52.60  11245.31
#> 27  56.34  87211.92
#> 28  51.66  19089.86
#> 29  55.60  39296.56
#> 32  56.87  27397.94
#> 34  51.75  33282.42
#> 36  58.20  42716.78
#> 37  45.97  14750.82
#> 40  45.36  37893.85
#> 41  50.82  79715.65
#> 43  51.25  33533.14
#> 44  55.70  11754.81
#> 45  45.72  53926.92
#> 46  54.18  62439.00
#> 48  47.59  39717.18
#> 50  43.41  26871.96
#> 58  41.49  17167.22
#> 60  47.26  18717.65
#> 61  44.95  23901.68
#> 65  52.45  22580.48
#> 66  46.86  18090.96
#> 72  43.20  39121.91
#> 75  47.26  32890.27
#> 79  53.91  21690.44
#> 81  47.38  19447.91
#> 85  48.25  70492.93
#> 89  53.41  68269.57
#> 90  39.34  27722.85
#> 94  38.42  77935.84
#> 95  49.25  21387.73
#> 96  46.59  27212.97
#> 98  52.30  39810.93
#> 100 63.05  39187.82
#> 101 51.94  33679.24
#> 103 49.73  51501.67
#> 104 50.61  16646.84
#> 105 48.65  33817.47
#> 106 57.46  25583.93
#> 109 43.36  12983.80
#> 111 41.76   8094.18
#> 113 53.32  11236.83
#> 114 46.79  70962.40
#> 115 49.39  19367.16
#> 116 45.20  21310.86
#> 118 41.39   9985.22
#> 119 40.64  30629.33
#> 121 48.35  55165.20
#> 122 56.06  12094.68
#> 125 53.78  32847.71
#> 126 53.36  52045.02
#> 127 48.83  40646.62
#> 129 55.98 100309.23
#> 130 51.38  20748.25
#> 131 54.08  34367.45
#> 133 41.45  19374.61
#> 134 54.08  25979.89
#> 136 47.46  46199.71
#> 137 50.89  49010.62
#> 140 55.88  23604.81
#> 142 46.24  27499.53
#> 146 45.42  26603.81
#> 147 40.17  29580.72
#> 150 43.35  50426.37
#> 151 39.30  29254.38
#> 152 45.04  28207.93
#> 157 52.68  44274.66
#> 160 50.47  26091.96
#> 161 46.42  18601.40
#> 164 39.77  63542.76
#> 165 43.85  17600.16
#> 166 44.11  16329.28
#> 167 44.42  20105.81
#> 168 50.87  16242.56
#> 170 41.13  61729.89
#> 175 40.60  18596.51
#> 181 51.74  42808.07
#> 182 50.66  60793.71
#> 185 58.47  76337.07
#> 186 53.86 107635.28
#> 188 56.12  54055.36
#> 189 55.24  42056.63
#> 190 59.70  38138.41
#> 191 58.68  28378.73
#> 192 53.63  65632.59
#> 193 56.87  43280.95
#> 195 56.57  10786.18
#> 196 54.03 102447.86
#> 197 55.36  24780.15
#> 200 49.73  45915.40
#> 203 52.83  24063.40
#> 205 42.81  14850.67
#> 207 53.59  45711.19
#> 212 50.80  54805.46
#> 213 48.80  12551.52
#> 214 52.60  38143.06
#> 217 49.73  52900.84
#> 218 45.97  27764.07
#> 219 46.10  62770.31
#> 220 57.05  25304.64
#> 221 52.88   9630.01
#> 223 57.99  31329.58
#> 225 49.89  37576.85
#> 228 47.58  33966.17
#> 229 50.84  57896.25
#> 230 45.07  94716.05
#> 231 55.81   8830.13
#> 232 55.28  54484.01
#> 233 47.62  22514.64
#> 238 49.84  51441.39
#> 239 58.06  19587.51
#> 248 49.37  29933.19
#> 249 47.62  20069.90
#> 250 48.28  33298.14
#> 252 46.13  44607.27
#> 253 46.55  18382.48
#> 255 46.92  77020.65
#> 256 52.46  42693.58
#> 264 42.15  73643.36
#> 266 46.85  30754.09
#> 267 53.93  30275.02
#> 273 49.73 123333.12
#> 274 39.60  24222.98
#> 280 52.37  20524.30
#> 287 50.78  25755.02
#> 288 50.40  14287.58
#> 291 54.30  54468.52
#> 299 50.81  22672.77
#> 300 46.99  26147.39
#> 304 53.85  30403.14
#> 306 48.43  44858.57
#> 308 41.11  16359.84
#> 313 53.03  22043.11
#> 315 46.17  46157.61
#> 318 57.43  71287.41
#> 322 38.34  85702.58
#> 323 51.72  53500.62
#> 325 49.14  32824.50
#> 329 48.45  16971.35
#> 331 48.91  13682.47
#> 335 42.52  14063.28
#> 337 57.01  34796.06
#> 339 53.07  55444.58
#> 341 53.31  86681.35
#> 343 42.96  32830.35
#> 344 47.36  34054.38
#> 346 43.90  16422.64
#> 347 51.78  17602.48
#> 348 51.56  30993.50
#> 350 51.28  34288.24
#> 353 44.52  11196.80
#> 354 52.24  17453.28
#> 358 45.44  23682.33
#> 361 49.42  30102.37
#> 362 52.87  56050.31
#> 364 57.17  75097.22
#> 366 65.34  43754.02
#> 376 52.11  40420.66
#> 378 42.94  64739.48
#> 379 51.36  24660.54
#> 380 46.65  20138.78
#> 383 44.60  17265.20
#> 388 51.40  52539.00
#> 389 59.34  50188.65
#> 393 48.31  30155.84
#> 394 54.73  19302.64
#> 395 51.08  21853.41
#> 397 62.49 106767.55
#> 400 56.94  39058.01
#> 401 64.38 118665.14
#> 402 53.68  21194.29
#> 404 54.07  29406.38
#> 405 49.34  15751.77
#> 406 54.61  34625.42
#> 407 57.58  22369.22
#> 409 52.55  74039.67
#> 413 56.98  22329.22
#> 415 54.23  30067.98
#> 416 51.49  50394.21
#> 418 55.93  41686.86
#> 420 53.39  46922.51
#> 421 53.96  32458.87
#> 422 49.78  71746.65
#> 424 51.80  50832.60
#> 425 55.11  77793.77
#> 426 52.63  56512.26
#> 427 49.90  42502.70
#> 428 46.36  20272.14
#> 429 54.28  49516.70
#> 432 54.79  27754.58
#> 434 48.48  18269.79
#> 438 47.91  23819.27
#> 440 46.61 129579.06
#> 442 46.11  43455.74
#> 443 52.95  18852.39
#> 444 29.99  15113.48
#> 445 43.78  66136.51
#> 447 41.29  27089.57
#> 449 56.40  44117.75
#> 450 50.98  35275.99
#> 453 44.76  25647.10
#> 455 49.97  65138.66
#> 461 48.60  18700.17
#> 464 55.63  20127.22
#> 466 48.69  17497.64
#> 468 48.27  29198.73
#> 469 50.28  25843.26
#> 472 36.49  43061.76
#> 473 52.25  17412.54
#> 474 46.55  20890.76
#> 478 56.13   8237.33
#> 484 52.92  30656.66
#> 495 52.88  75586.98
#> 496 55.09   9815.00
#> 499 50.69  30083.74
#> 501 58.43  32522.75
#> 504 47.32  27693.68
#> 505 51.44  25016.55
#> 506 48.80  11391.82
#> 511 45.42  25932.62
#> 513 44.69  28530.12
#> 516 40.18  37693.18
#> 517 47.73  30789.89
#> 519 43.33  31519.74
#> 525 47.95  43895.90
#> 527 57.35  19791.18
#> 528 39.87  32579.26
#> 530 53.62  62074.03
#> 531 41.98  31881.69
#> 532 52.29  51536.65
#> 534 51.00  20570.68
#> 536 52.06  30950.65
#> 538 55.01  25151.38
#> 539 51.17  43643.24
#> 540 42.08  71250.62
#> 543 44.67  10210.56
#> 545 48.29  40615.79
#> 546 46.59  20448.08
#> 548 44.83  14182.16
#> 550 48.54  54287.41
#> 552 44.40  46442.89
#> 554 47.17  16036.48
#> 555 44.78  48628.07
#> 556 43.67  54834.64
#> 557 50.31  24120.45
#> 559 31.31  28964.58
#> 560 44.13  12939.45
#> 564 46.95  24690.51
#> 565 54.14  11732.99
#> 566 45.06  24339.63
#> 570 54.18  21000.04
#> 571 56.83  19150.29
#> 573 41.83  10364.57
#> 574 57.25  83721.54
#> 575 52.61  16465.66
#> 576 49.61  14986.53
#> 579 55.41  15250.23
#> 583 36.31  23000.57
#> 584 47.21  54584.30
#> 586 47.12  32297.47
#> 592 62.03 144764.24
#> 593 65.33  62929.93
#> 594 56.73  47548.45
#> 595 62.59  52190.64
#> 596 52.19  47092.76
#> 601 48.51   9353.98
#> 603 47.12  66728.13
#> 605 56.94  27973.46
#> 606 50.47  37448.91
#> 609 49.40  33888.20
#> 611 50.82  35740.05
#> 613 53.51  24667.10
#> 620 47.29  20558.70
#> 621 43.42  10296.57
#> 623 55.21  11819.52
#> 626 51.75  28483.56
#> 629 43.56  36133.09
#> 631 49.41  39974.83
#> 633 46.89  55514.67
#> 635 40.23  17219.22
#> 645 54.16  57390.41
#> 650 60.27  64260.62
#> 654 50.43  53652.65
#> 655 51.07  11419.79
#> 656 55.16  22722.24
#> 661 46.55  39042.66
#> 662 57.13  66495.43
#> 664 44.38  25683.58
#> 665 48.23  21610.66
#> 666 50.13  67466.70
#> 667 55.21  21844.88
#> 669 56.23  37220.65
#> 675 60.44  34264.84
#> 682 42.06 182886.22
#> 687 52.71  33477.46
#> 688 48.52  69387.51
#> 689 60.35  31338.43
#> 690 51.62  84940.45
#> 692 49.97  38267.69
#> 698 50.65  37791.37
#> 700 53.96  82979.07
#> 703 43.09  11499.55
#> 704 48.26  19768.24
#> 707 47.09  12059.49
#> 709 55.35  20114.64
#> 710 55.16  38289.64
#> 714 53.58  54495.55
#> 715 59.36  22237.98
#> 717 46.90  14617.62
#> 718 56.52  39492.98
#> 719 47.03  22586.48
#> 723 54.98  20042.68
#> 725 49.87  41435.15
#> 727 50.46  52638.66
#> 728 51.63   9911.82
#> 730 62.31  62781.85
#> 732 46.05  49712.79
#> 733 51.87  12590.63
#> 734 45.08  27615.51
#> 738 50.14  57255.41
#> 742 49.07  33600.88
#> 744 42.42  61737.25
#> 747 48.19  30563.47
#> 751 57.84  33307.25
#> 752 55.75  55927.38
#> 755 52.47  29695.26
#> 757 58.10  20191.67
#> 758 53.70  19345.38
#> 760 57.68  56191.83
#> 762 44.01  45443.21
#> 765 59.52  23540.91
#> 766 61.30  39284.40
#> 769 49.83  54943.24
#> 770 55.79  41379.83
#> 771 50.05  31861.92
#> 774 42.55  40431.32
#> 776 46.88  22352.76
#> 777 44.30  14168.37
#> 781 47.82  18626.49
#> 784 46.26  62579.79
#> 785 51.37  41814.17
#> 788 43.81  19969.46
#> 789 58.37  25593.61
#> 790 47.46  47322.40
#> 791 50.34  50133.02
#> 792 53.95   8967.47
#> 794 44.76  49022.23
#> 795 47.09  31955.73
#> 799 49.68  20470.53
#> 800 49.93  39199.82
#> 802 45.42  60414.30
#> 805 45.94  17140.81
#> 807 40.41  35744.19
#> 809 54.08  20261.38
#> 811 52.74  22741.51
#> 818 52.48  26611.77
#> 819 51.73  28413.51
#> 822 47.75  30827.69
#> 823 53.80  47225.48
#> 824 45.85  84903.16
#> 828 46.29  70466.56
#> 830 39.52  25511.95
#> 832 46.63  26933.14
#> 847 56.38  26778.58
#> 850 55.69  18081.47
#> 852 53.85  58491.62
#> 853 46.58  35250.68
#> 855 45.96  57277.74
#> 859 42.91  28930.46
#> 861 51.97  11619.26
#> 862 47.19  32151.69
#> 864 43.96  46455.99
#> 868 48.44  41093.24
#> 871 56.78  25238.75
#> 874 42.81  18813.99
#> 875 53.92  66167.56
#> 877 60.24  21331.66
#> 880 43.66  17767.68
#> 884 45.63  39015.78
#> 885 43.57  32084.39
#> 886 47.37  31505.23
#> 887 46.24  18611.43
#> 888 42.37  26939.35
#> 890 59.59  28705.72
#> 891 57.40  44558.32
#> 892 55.52  23029.45
#> 894 60.12  19097.21
#> 898 62.91 143381.52
#> 899 52.41  62092.21
ilm_subset(d, subset = ilm_sample(10, seed = 1), cols = c("id", "date"),
           cols_negate = TRUE)
#>       grp  site  flag consented score    income visits claims downtime cohort
#> 129 alpha North FALSE      TRUE 55.98 100309.23     10      4        0   2024
#> 187 gamma South  TRUE      TRUE 63.91  28676.24      1     27        0   2024
#> 270  beta  East  TRUE      TRUE 41.99  30287.37      4      0        0   2024
#> 299 alpha South FALSE      TRUE 50.81  22672.77      8      6        4   2024
#> 307  beta South  TRUE      TRUE 42.91  10337.46      1      1        0   2024
#> 471  beta North FALSE      TRUE 53.53  22711.16      5      2        2   2024
#> 509 gamma  East  TRUE      TRUE 47.75  20570.32      3      4        7   2024
#> 597  beta  East FALSE      TRUE 54.96  93063.82      3     10        4   2024
#> 679  beta  East FALSE      TRUE 63.65  55891.18      3      7        0   2024
#> 836 gamma  East FALSE      TRUE 45.05  90323.58      2      1        1   2024
#>     lab_value
#> 129        NA
#> 187      8.59
#> 270      5.10
#> 299      7.96
#> 307      6.36
#> 471      7.88
#> 509      6.93
#> 597      6.87
#> 679      4.28
#> 836      6.47